app/alembic/versions/2025_12_30_1550-604941279867_initial_migration.py
"""initial_migration

Revision ID: 604941279867
Revises: 
Create Date: 2025-12-30 15:50:41.465405

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa
import sqlmodel



# revision identifiers, used by Alembic.
revision: str = '604941279867'
down_revision: Union[str, Sequence[str], None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    """Upgrade schema."""
    pass


def downgrade() -> None:
    """Downgrade schema."""
    pass
