pyproject.toml
[project]
name = "langchain-text-analyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "black>=25.12.0",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
